Increases Item Number, COS ID, and Object ID
Increases the Item Number, COS ID, and Object ID maximum values up to the 32 bit integer limit values to prevent overflow errors so that users can freely avoid conflicts without extracting and editing the databases manually. While some values, like the COS ID, may support only up to 500, there are various resources that inform the users of that and is only meant as a precaution for when ID limit patches are created.