Adds Custom Head template, QOL improvements
Adds a template for Custom Head models to the chritm editor, and improves the tool tips for the editors to add more information about how they should be formatted. Also changed the default sort index for modules to 800 that way they can appear before t-shirt modules, but you can have them appear behind the t-shirts as normal by setting it to an ID above 905.